Fixed figure reporting C

+++ example/ping_pong.c
@@ -18,8 +18,8 @@ double get_wall_time(){
 
 int main(int argc, char** argv) {
   const int root = 0;
-  int payload;
-  int iterations;
+  long payload;
+  long iterations;
 
   MPI_Init(&argc,&argv);
   int rank;
@@ -52,8 +52,8 @@ int main(int argc, char** argv) {
     }
   }
 
-  MPI_Bcast(&iterations, 1, MPI_INT, root, MPI_COMM_WORLD);
-  MPI_Bcast(&payload, 1, MPI_INT, root, MPI_COMM_WORLD);
+  MPI_Bcast(&iterations, 1, MPI_LONG, root, MPI_COMM_WORLD);
+  MPI_Bcast(&payload, 1, MPI_LONG, root, MPI_COMM_WORLD);
   double* vector = malloc(payload * sizeof(double));
 
   for (int i = 0; i < payload; i++) {
